change outline for calendar
