+++ app/src/main/res/layout/activity_main.xml
@@ -1,132 +1,22 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.drawerlayout.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    tools:context=".MainActivity">
 
-    <androidx.constraintlayout.widget.ConstraintLayout
+
+    <androidx.fragment.app.FragmentContainerView
+        android:id="@+id/fragment_container_main"
+        android:name="androidx.navigation.fragment.NavHostFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MainActivity">
+        app:defaultNavHost="true"
+        app:navGraph="@navigation/nav_graph"
 
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-
-            android:layout_marginTop="16dp"
-            android:text="@string/hangeul_oneul"
-            android:textSize="46sp"
-            android:textColor="@color/black"
-
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <androidx.appcompat.widget.AppCompatImageButton
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="28dp"
-            android:layout_marginBottom="12dp"
-
-            android:background="@color/white"
-            android:src="@drawable/ic_menu"
-            app:layout_constraintBottom_toTopOf="@id/calendar_conatainer_main"
-            app:layout_constraintStart_toStartOf="parent" />
-
-        <androidx.appcompat.widget.AppCompatToggleButton
-            android:layout_width="42sp"
-            android:layout_height="21sp"
-            android:layout_marginEnd="28dp"
-            android:layout_marginBottom="12dp"
-            android:background="@drawable/toggle_selector"
-
-            android:textOff=" "
-            android:textOn=" "
-            app:layout_constraintBottom_toTopOf="@id/calendar_conatainer_main"
-            app:layout_constraintEnd_toEndOf="parent" />
-
-        <LinearLayout
-            android:id="@+id/calendar_conatainer_main"
-            android:layout_width="match_parent"
-            android:layout_height="480sp"
-            android:orientation="vertical"
-
-            android:layout_margin="28dp"
-            android:background="@drawable/ic_calendar_box"
-
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            >
-
-            <RelativeLayout
-                android:layout_width="62sp"
-                android:layout_height="32sp"
-                android:background="@drawable/ic_small_button"
-                android:layout_gravity="center_horizontal"
-                android:layout_margin="12dp"
-                >
-                <TextView
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="내 캘린더"
-                    android:textSize="18dp"
-                    android:textColor="@color/primary"
-                    android:layout_centerInParent="true"
-                    />
-            </RelativeLayout>
+        />
 
 
-            <androidx.fragment.app.FragmentContainerView
-                android:id="@+id/fragment_container_calendar_main"
-                android:name="androidx.navigation.fragment.NavHostFragment"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                app:defaultNavHost="true"
-                app:navGraph="@navigation/nav_graph"
-
-                />
-        </LinearLayout>
-
-        <RelativeLayout
-            android:id="@+id/button_add_event_main"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-
-            android:layout_marginBottom="16dp"
-            android:background="@drawable/ic_add_buttons"
-            android:clickable="true"
-            app:layout_constraintBottom_toBottomOf="parent"
-
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent">
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_centerInParent="true"
-                android:text="일정추가"
-                android:textColor="@color/primary"
-                android:textSize="36sp" />
-
-        </RelativeLayout>
-
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-    <com.google.android.material.navigation.NavigationView
-
-        android:layout_width="wrap_content"
-        android:layout_height="match_parent"
-
-        android:fitsSystemWindows="true"
-        android:visibility="invisible"
-
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:menu="@menu/navigation_drawer" />
-
-</androidx.drawerlayout.widget.DrawerLayout>
+</androidx.constraintlayout.widget.ConstraintLayout>
