+++ app/src/main/res/layout/fragment_schedule_calender.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical">
+
+   <com.prolificinteractive.materialcalendarview.MaterialCalendarView
+       xmlns:app="http://schemas.android.com/apk/res-auto"
+       android:id="@+id/schedule_calendarView"
+       android:layout_width="match_parent"
+       android:layout_height="match_parent"
+       app:mcv_showOtherDates="defaults"
+       app:mcv_selectionColor="@color/primary"
+       app:mcv_rightArrow="@drawable/ic_button_circle_arrow_right"
+       app:mcv_leftArrow="@drawable/ic_button_circle_arrow_left"
+      />
+
+
+</LinearLayout>
